LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/transform-percent-009-expected.txt


FAIL Percentage 'translate(...)' with 'fill-box' transform-box assert_equals: expected "matrix(1, 0, 0, 1, 50, 50)" but got "matrix(1, 0, 0, 1, 0, 0)"
FAIL Percentage 'translate(...)' with 'view-box' transform-box assert_equals: expected "matrix(1, 0, 0, 1, 150, 75)" but got "matrix(1, 0, 0, 1, 0, 0)"
FAIL Percentage 'translate(...)' with 'view-box' transform-box in nested <svg> with 'viewBox' assert_equals: expected "matrix(1, 0, 0, 1, 25, 25)" but got "matrix(1, 0, 0, 1, 0, 0)"
FAIL Percentage 'translate(...)' with 'view-box' transform-box in nested <svg> without 'viewBox' assert_equals: expected "matrix(1, 0, 0, 1, 50, 50)" but got "matrix(1, 0, 0, 1, 0, 0)"
